move TrieLog from datatypes to plugin-api
