+++ plugin-api/build.gradle
@@ -69,7 +69,7 @@ Calculated : ${currentHash}
 tasks.register('checkAPIChanges', FileStateChecker) {
   description = "Checks that the API for the Plugin-API project does not change without deliberate thought"
   files = sourceSets.main.allJava.files
-  knownHash = 'nQ4sgzgeGXts5YKfNUQN0urTXtGGk2QfzB9V6BDNAYc='
+  knownHash = 'YmNahYbhpknZicR+0aAUNn7U33Eqs1VcA+H0IuFCXi0='
 }
 check.dependsOn('checkAPIChanges')
